Do Annika's suggestions, move out of own file
